backend/requirements.txt
Flask==2.3.3
twilio==8.10.0
pandas>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
openai>=1.3.0
anthropic>=0.7.0